# ai_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "온식구 AI 서버",
    "description": "가족 유대감을 위한 AI 질문 생성 서버",
    "version": "2.1.0"
  },
  "paths": {
...
    "/api/v1/questions/instance/langchain": {
      "post": {
        "tags": ["기본 질문"],
        "summary": "(준비중) LangChain 기반 질문 생성",
        "description": "LangChain 구현은 준비 중입니다.",
        "operationId": "create_question_instance_langchain_api_v1_questions_instance_langchain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionInstanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/assign": {
      "post": {
        "tags": ["기본 질문"],
        "summary": "멤버 할당",
        "description": "최근 30회 기준으로 덜 받은 멤버에 확률을 부여해 pick_count명 선정",
        "operationId": "assign_members_api_v1_questions_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberAssignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmotionScores": {
        "type": "object",
        "title": "EmotionScores",
        "properties": {
          "joy": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "sadness": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "anger": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "fear": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "neutral": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"}
        }
      },
      "AnalysisScores": {
        "type": "object",
        "title": "AnalysisScores",
        "properties": {
          "sentiment": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "-1.0~1.0"},
          "emotion": {"anyOf": [{"$ref": "#/components/schemas/EmotionScores"}, {"type": "null"}], "description": "감정 스코어"},
          "relevance_to_question": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "relevance_to_category": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "toxicity": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "0~1"},
          "length": {"anyOf": [{"type": "integer"}, {"type": "null"}], "description": "답변 길이"}
        }
      },
      "AnswerAnalysisHint": {
        "type": "object",
        "title": "AnswerAnalysisHint",
        "properties": {
          "summary": {"anyOf": [{"type": "string"}, {"type": "null"}]},
          "categories": {"anyOf": [{"type": "array", "items": {"type": "string"}}, {"type": "null"}]},
          "scores": {"anyOf": [{"$ref": "#/components/schemas/AnalysisScores"}, {"type": "null"}]},
          "keywords": {"anyOf": [{"type": "array", "items": {"type": "string"}}, {"type": "null"}]}
        }
      },
...
      "QuestionGenerateRequest": {
        "type": "object",
        "title": "QuestionGenerateRequest",
        "required": ["content"],
        "properties": {
          "content": {"type": "string", "title": "Content", "description": "질문 생성의 기반이 되는 베이스 문구"},
          "language": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Language", "description": "언어 코드(기본 ko)", "default": "ko"},
          "tone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Tone", "description": "톤/스타일"},
          "category": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category", "description": "카테고리"},
          "tags": {"anyOf": [{"type": "array", "items": {"type": "string"}}, {"type": "null"}], "title": "Tags", "description": "태그 목록"},
          "subject_required": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Subject Required", "description": "주제 인물 필요 여부", "default": false},
          "mood": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Mood", "description": "원하는 분위기(선택)"},
          "answer_analysis": {"anyOf": [{"$ref": "#/components/schemas/AnswerAnalysisHint"}, {"type": "null"}], "title": "Answer Analysis", "description": "답변 분석 힌트"}
        }
      },
      "QuestionInstanceResponse": {
        "type": "object",
        "title": "QuestionInstanceResponse",
        "required": ["content", "generated_by", "generation_model", "generation_parameters", "generation_prompt", "generation_metadata", "generation_confidence", "generated_at"],
        "properties": {
          "content": {"type": "string", "title": "Content", "description": "생성된 질문 내용"},
          "planned_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Planned Date", "description": "계획 날짜"},
          "status": {"type": "string", "title": "Status", "description": "상태: draft|scheduled|sent 등", "default": "draft"},
          "generated_by": {"type": "string", "title": "Generated By", "description": "생성 주체: ai|manual"},
          "generation_model": {"type": "string", "title": "Generation Model", "description": "사용 모델명"},
          "generation_parameters": {"type": "object", "title": "Generation Parameters", "description": "모델 호출 파라미터"},
          "generation_prompt": {"type": "string", "title": "Generation Prompt", "description": "프롬프트 원문"},
          "generation_metadata": {"type": "object", "title": "Generation Metadata", "description": "생성 메타데이터"},
          "generation_confidence": {"type": "number", "title": "Generation Confidence", "description": "생성 신뢰도(0~1)"},
          "generated_at": {"type": "string", "format": "date-time", "title": "Generated At", "description": "생성 시각"}
        }
      },
      "AssignMemberInput": {
        "type": "object",
        "title": "AssignMemberInput",
        "required": ["member_id", "assigned_count_30"],
        "properties": {
          "member_id": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Member Id", "description": "멤버 ID"},
          "assigned_count_30": {"type": "integer", "minimum": 0, "title": "Assigned Count 30", "description": "최근 30회 내 할당 횟수(0 이상)"}
        }
      },
      "AssignOptions": {
        "type": "object",
        "title": "AssignOptions",
        "properties": {
          "epsilon": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Epsilon", "description": "가중치 하한값", "default": 1e-9}
        }
      },
      "MemberAssignRequest": {
        "type": "object",
        "title": "MemberAssignRequest",
        "required": ["family_id", "members", "pick_count"],
        "properties": {
          "family_id": {"anyOf": [{"type": "integer"}, {"type": "string"}], "title": "Family Id", "description": "가족 ID"},
          "members": {"type": "array", "items": {"$ref": "#/components/schemas/AssignMemberInput"}, "title": "Members", "description": "후보 멤버 목록"},
          "pick_count": {"type": "integer", "minimum": 1, "title": "Pick Count", "description": "선정할 멤버 수"},
          "options": {"anyOf": [{"$ref": "#/components/schemas/AssignOptions"}, {"type": "null"}], "description": "선정 옵션"}
        }
      },
      "MemberAssignResponse": {
        "type": "object",
        "title": "MemberAssignResponse",
        "required": ["member_ids"],
        "properties": {
          "member_ids": {"type": "array", "items": {"anyOf": [{"type": "integer"}, {"type": "string"}]}, "title": "Member Ids", "description": "선정된 멤버 ID 목록"},
          "version": {"type": "string", "title": "Version", "description": "로직 버전", "default": "assign-v1"}
        }
      },
      "HTTPValidationError": {
        "type": "object",
        "title": "HTTPValidationError",
        "properties": {
          "detail": {"type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}, "title": "Detail"}
        }
      },
      "ValidationError": {
        "type": "object",
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "properties": {
          "loc": {"type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "title": "Location"},
          "msg": {"type": "string", "title": "Message"},
          "type": {"type": "string", "title": "Error Type"}
        }
      }
    }
  }
}